refactor: イベント詳細表示用の重複クエリを整理する
- EventShowResources concern を追加
- EventsController と EventPreferencesController の重複処理を共通化
- show 固有の event_preference は controller 側に残して責務を整理
